Do not crash when the directory doesn't exist.
Fix regression introduced in r125670.

TBR=nsylvain@chromium.org
BUG=
TEST=


Review URL: http://codereview.chromium.org/9651003

git-svn-id: svn://svn.chromium.org/chrome/trunk/tools/build@125673 0039d316-1c4b-4281-b951-d872f2087c98

+++ scripts/slave/cleanup_temp.py
@@ -105,10 +105,11 @@ def main_win():
   # 'process build' step.
   crash_reports = os.path.join(
       os.environ['LOCALAPPDATA'], 'Chromium', 'User Data', 'Crash Reports')
-  for filename in os.listdir(crash_reports):
-    filepath = os.path.join(crash_reports, filename)
-    if os.path.isfile(filepath):
-      os.remove(filepath)
+  if os.path.isdir(crash_reports):
+    for filename in os.listdir(crash_reports):
+      filepath = os.path.join(crash_reports, filename)
+      if os.path.isfile(filepath):
+        os.remove(filepath)
   return 0
